+++ package.json
@@ -10,6 +10,8 @@
   },
   "dependencies": {
     "@react-navigation/bottom-tabs": "^6.3.1",
+    "@react-navigation/native": "^6.0.10",
+    "@react-navigation/native-stack": "^6.6.1",
     "@types/styled-components": "^5.1.24",
     "expo": "~44.0.2",
     "expo-splash-screen": "~0.14.1",
@@ -20,6 +22,8 @@
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
     "react-native-paper": "^4.12.0",
+    "react-native-safe-area-context": "^4.2.4",
+    "react-native-screens": "^3.13.1",
     "react-native-vector-icons": "^9.1.0",
     "react-native-web": "0.17.1",
     "styled-components": "^5.3.5"
