+++ src/Pages/Signin/index.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Signin() {
+  return <div>Signin</div>;
+}
